+++ test/unit-test/test_bootloader_upgrade_check.c
@@ -7,6 +7,7 @@
 #include <string.h>
 #include <cmocka.h>
 
+#include "bootloader/bootloader_product.h"
 #include "bootloader/stage0/stage0_descriptor.h"
 #include "bootloader_upgrade/bootloader_upgrade.h"
 #include "bootloader_upgrade/firmware_installer_check.h"